refactor: 칸반보드에서 섹션 분리
